#issue108: implemented courseController instead of overLoading userController, courseControl the flow for course related routes.
